+++ .gitignore
@@ -3,6 +3,7 @@
 ~*
 **/*.csv
 *.yaml
+*.yml
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
